Blogilo: Move Blogger API implementation from KBlog to Blogilo
The Blogger API implementation requires LibKGAPI, but LibKGAPI depends
on kdepimlibs. To avoid circular dependency, the Blogger API is temporarily
moved from kdepimlibs' KBlog directly to Blogilo, which requires the
Blogger API.

The code can be moved back to KBlog in Frameworks 5, when the dependency
circle between LibKGAPI and kdepimlibs frameworks will be broken.

REVIEW: 119196